% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSpatial.R
\name{LoadSpatial}
\alias{LoadSpatial}
\title{Loading data from a Seurat object or a data frame.}
\usage{
LoadSpatial(InputData, Dimension = 2)
}
\arguments{
\item{InputData}{A Seurat spatial object or a M x (D + N) data matrix representing the D-dimensional coordinates and expressions of N genes on M spots. The coordinates should be placed at the first D columns}

\item{Dimension}{The dimension of coordinates}
}
\value{
A list of two data frame:
\item{Coords}{A M x D matrix representing D-dimensional coordinates for M spots}
\item{ExpMatrix}{A sparse, N x M expression matrix in dgCMatrix class with N genes and M spots}
}
\description{
A function to load and filter data from a Seurat object or a data frame.
}
